+++ GeoTag/META-INF/MANIFEST.MF
@@ -1,7 +1,7 @@
 Manifest-Version: 1.0
 Bundle-SymbolicName: GeoTag;singleton:=true
 Bundle-Name: GeoTag
-Bundle-Version: 1.3.1
+Bundle-Version: 1.3.2
 Require-Bundle: org.eclipse.jface.databinding;bundle-version="1.7.0",
  javax.inject;bundle-version="0.0.0",
  org.eclipse.core.runtime;bundle-version="0.0.0",
