+++ GeoTag/src/main/java/joachimeichborn/geotag/misc/ColorGenerator.java
@@ -17,7 +17,7 @@ You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-package joachimeichborn.geotag.utils;
+package joachimeichborn.geotag.misc;
 
 import java.util.ArrayList;
 import java.util.List;
